backend/requirements.txt
# Backend Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# MQTT
paho-mqtt==1.6.1

# Data Processing
numpy==1.24.3
scipy==1.11.4

# Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# WebSocket
websockets==12.0

# CORS
python-dotenv==1.0.0

# Development
pytest==7.4.3
httpx==0.25.2